Refactor and completed navbar
