+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.marill"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,8 +22,8 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".RegistrationActivity" />
-        <activity android:name=".HomePageActivity" />
+        <activity android:name=".activities.RegistrationActivity" />
+        <activity android:name=".activities.HomePageActivity" />
     </application>
 
 </manifest>
