+++ app/src/main/java/com/example/marill_many_events/NavbarListener.java
@@ -1,5 +1,5 @@
 package com.example.marill_many_events;
 
 public interface NavbarListener {
-    void onProfileSelected(String deviceId); // Method to handle profile selection
+    void onProfileSelected(); // Method to handle profile selection
 }
