+++ app/src/main/java/com/example/marill_many_events/activities/HomePageActivity.java
@@ -1,29 +1,26 @@
-package com.example.marill_many_events;
+package com.example.marill_many_events.activities;
 
 import static com.google.firebase.appcheck.internal.util.Logger.TAG;
 
 import android.os.Bundle;
 import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.MenuItem;
-import android.view.View;
-import android.view.ViewGroup;
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-import com.google.android.material.bottomnavigation.BottomNavigationView;
 
-import android.os.Bundle;
+import com.example.marill_many_events.fragments.NavbarFragment;
+import com.example.marill_many_events.NavbarListener;
+import com.example.marill_many_events.R;
+import com.example.marill_many_events.fragments.RegistrationFragment;
+
 import android.widget.FrameLayout;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.Fragment;
+
+import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 
 public class HomePageActivity extends AppCompatActivity implements NavbarListener {
 
-    private FirebaseFirestore db;
+    private FirebaseFirestore firestore;
+    private CollectionReference usersRef;
     private String deviceId; // Store deviceId here
 
     @Override
@@ -32,29 +29,23 @@ public class HomePageActivity extends AppCompatActivity implements NavbarListene
         setContentView(R.layout.activity_home);
         deviceId = getIntent().getStringExtra("deviceId"); // Retrieve deviceId
 
-        db = FirebaseFirestore.getInstance();
-
-        PlaceholderFragment placeholderFragment = new PlaceholderFragment();
-        getSupportFragmentManager().beginTransaction()
-                .replace(R.id.fragment_container, placeholderFragment)
-                .commit();
+        firestore = FirebaseFirestore.getInstance();
+        usersRef = firestore.collection("users");
 
         Log.d("HomePageActivity", "Fragment Container Visibility: " + findViewById(R.id.fragment_container).getVisibility());
 
 
-
-
-        // Set up Navbar
-        Navbar navbar = new Navbar();
+        // Set up NavbarFragment
+        NavbarFragment navbarFragment = new NavbarFragment();
         getSupportFragmentManager().beginTransaction()
-                .replace(R.id.navbar_container, navbar)
+                .replace(R.id.navbar_container, navbarFragment)
                 .commit();
 
 
     }
 
     @Override
-    public void onProfileSelected(String deviceId) {
+    public void onProfileSelected() {
         // Open the RegistrationFragment when profile is selected
         RegistrationFragment registrationFragment = new RegistrationFragment();
 
@@ -75,7 +66,7 @@ public class HomePageActivity extends AppCompatActivity implements NavbarListene
         FrameLayout fragmentContainer = findViewById(R.id.fragment_container);
         FrameLayout navbarContainer = findViewById(R.id.navbar_container);
         Log.d(TAG, "Fragment Container Visibility: " + fragmentContainer.getVisibility());
-        Log.d(TAG, "Navbar Container Visibility: " + navbarContainer.getVisibility());
+        Log.d(TAG, "NavbarFragment Container Visibility: " + navbarContainer.getVisibility());
 
 
     }
