+++ app/src/main/java/com/example/marill_many_events/activities/MainActivity.java
@@ -1,35 +1,21 @@
-package com.example.marill_many_events;
+package com.example.marill_many_events.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.widget.Button;
 import android.provider.Settings;
 
-import androidx.activity.EdgeToEdge;
 import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
 import androidx.activity.result.ActivityResultLauncher;
 
-import android.content.Intent;
-import android.os.Bundle;
-import android.provider.Settings;
 import android.util.Log;
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import android.content.Intent;
-import android.os.Bundle;
-import android.util.Log;
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
+
 import androidx.fragment.app.Fragment;
 
-import com.google.firebase.FirebaseApp;
-import com.google.firebase.firestore.DocumentReference;
+import com.example.marill_many_events.R;
+import com.example.marill_many_events.fragments.RegistrationFragment;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.QuerySnapshot;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.CollectionReference;
