+++ app/src/main/res/values/colors.xml
@@ -3,6 +3,6 @@
     <color name="blue_start">#FF006494</color>
     <color name="blue_end">#FF051923</color>
     <color name="light_blue">#FFDEEBF0</color>
-    <color name="navbar">#FF025177</color>
+    <color name="navbarFragment">#FF025177</color>
     <color name="grey">#FF373737</color>
 </resources>
